requirements.txt
nltk==3.8.1
numpy==1.24.2
openpyxl==3.1.0
pandas==1.5.3
pdfminer.six==20221105